fcl: Split fclspace.h into header and implementation.
