+++ plugins/fclrave/fclspace.h
@@ -26,39 +26,27 @@ typedef std::pair<Transform, CollisionObjectPtr> TransformCollisionPair;
 
 // Helper functions for conversions from OpenRAVE to FCL
 
-Vector ConvertVectorFromFCL(fcl::Vec3f const &v)
+inline Vector ConvertVectorFromFCL(fcl::Vec3f const &v)
 {
     return Vector(v[0], v[1], v[2]);
 }
 
-fcl::Vec3f ConvertVectorToFCL(Vector const &v)
+inline fcl::Vec3f ConvertVectorToFCL(Vector const &v)
 {
     return fcl::Vec3f(v.x, v.y, v.z);
 }
-fcl::Quaternion3f ConvertQuaternionToFCL(Vector const &v)
+inline fcl::Quaternion3f ConvertQuaternionToFCL(Vector const &v)
 {
     return fcl::Quaternion3f(v[0], v[1], v[2], v[3]);
 }
 
-Vector ConvertQuaternionFromFCL(fcl::Quaternion3f const &v) {
+inline Vector ConvertQuaternionFromFCL(fcl::Quaternion3f const &v) {
     return Vector(v.getW(), v.getX(), v.getY(), v.getZ());
 }
 
-fcl::AABB ConvertAABBToFcl(const OpenRAVE::AABB& bv) {
+inline fcl::AABB ConvertAABBToFcl(const OpenRAVE::AABB& bv) {
     return fcl::AABB(fcl::AABB(ConvertVectorToFCL(bv.pos)), ConvertVectorToFCL(bv.extents));
 }
-
-
-template <class T>
-CollisionGeometryPtr ConvertMeshToFCL(std::vector<fcl::Vec3f> const &points,std::vector<fcl::Triangle> const &triangles)
-{
-    std::shared_ptr< fcl::BVHModel<T> > const model = make_shared<fcl::BVHModel<T> >();
-    model->beginModel(triangles.size(), points.size());
-    model->addSubModel(points, triangles);
-    model->endModel();
-    return model;
-}
-
 
 /// \brief ensures vector size is at least size
 template <typename T>
@@ -87,11 +75,8 @@ public:
         class FCLGeometryInfo
         {
 public:
-            FCLGeometryInfo() : bFromKinBodyGeometry(false) {
-            }
-
-            FCLGeometryInfo(KinBody::GeometryPtr pgeom) : _pgeom(pgeom), bFromKinBodyGeometry(true) {
-            }
+            FCLGeometryInfo();
+            FCLGeometryInfo(KinBody::GeometryPtr pgeom);
 
             virtual ~FCLGeometryInfo() {
             }
@@ -108,10 +93,8 @@ public:
         class LinkInfo
         {
 public:
-            LinkInfo() : bFromKinBodyLink(false) {
-            }
-            LinkInfo(KinBody::LinkPtr plink) : _plink(plink), bFromKinBodyLink(true) {
-            }
+            LinkInfo();
+            LinkInfo(KinBody::LinkPtr plink);
 
             virtual ~LinkInfo() {
                 Reset();
@@ -153,24 +136,13 @@ public:
             bool bFromKinBodyLink; ///< if true, then from kinbodylink. Otherwise from standalone object that does not have any KinBody associations
         };
 
-        FCLKinBodyInfo() : nLastStamp(0), nLinkUpdateStamp(0), nGeometryUpdateStamp(0), nAttachedBodiesUpdateStamp(0), nActiveDOFUpdateStamp(0)
-        {
-        }
+        FCLKinBodyInfo();
 
         virtual ~FCLKinBodyInfo() {
             Reset();
         }
 
-        void Reset()
-        {
-            FOREACH(itlink, vlinks) {
-                (*itlink)->Reset();
-            }
-            vlinks.resize(0);
-            _geometrycallback.reset();
-            _geometrygroupcallback.reset();
-            _linkenablecallback.reset();
-        }
+        void Reset();
 
         KinBodyPtr GetBody()
         {
@@ -204,406 +176,43 @@ public:
     typedef boost::shared_ptr<FCLSpace::FCLKinBodyInfo::LinkInfo> LinkInfoPtr;
     typedef boost::function<void (FCLKinBodyInfoPtr)> SynchronizeCallbackFn;
 
-    FCLSpace(EnvironmentBasePtr penv, const std::string& userdatakey)
-        : _penv(penv), _userdatakey(userdatakey),
-        _currentpinfo(1, FCLKinBodyInfoPtr()), // initialize with one null pointer, this is a place holder for null pointer so that we can return by reference. env id 0 means invalid so it's consistent with the definition as well
-        _bIsSelfCollisionChecker(true)
-    {
-
-        // After many test, OBB seems to be the only real option (followed by kIOS which is needed for distance checking)
-        SetBVHRepresentation("OBB");
-    }
+    FCLSpace(EnvironmentBasePtr penv, const std::string& userdatakey);
 
     virtual ~FCLSpace()
     {
         DestroyEnvironment();
     }
 
-    void DestroyEnvironment()
-    {
-        RAVELOG_VERBOSE_FORMAT("destroying fcl collision environment (env %d) (userdatakey %s)", _penv->GetId()%_userdatakey);
-        for (KinBodyConstPtr pbody : _vecInitializedBodies) {
-            if (!pbody) {
-                continue;
-            }
-            FCLKinBodyInfoPtr& pinfo = GetInfo(*pbody);
-            if( !!pinfo ) {
-                pinfo->Reset();
-            }
-        }
-        // even after DestroyEnvironment is called, users of this class still try to access _currentpinfo
-        // in that case, null pointer should be returned, instead of range error from _currentpinfo.at(0). for that purpose, keep the first element here.
-        _currentpinfo.erase(_currentpinfo.begin() + 1, _currentpinfo.end());
-        _cachedpinfo.clear();
-        _vecInitializedBodies.clear();
-    }
+    void DestroyEnvironment();
 
-    FCLKinBodyInfoPtr InitKinBody(KinBodyConstPtr pbody, FCLKinBodyInfoPtr pinfo = FCLKinBodyInfoPtr(), bool bSetToCurrentPInfo=true)
-    {
-        if( !pinfo ) {
-            pinfo.reset(new FCLKinBodyInfo());
-            pinfo->_geometrygroup = _geometrygroup;
-        }
+    FCLKinBodyInfoPtr InitKinBody(KinBodyConstPtr pbody, FCLKinBodyInfoPtr pinfo = FCLKinBodyInfoPtr(), bool bSetToCurrentPInfo=true);
 
-        RAVELOG_VERBOSE_FORMAT("env=%s, self=%d, init body %s (%d)", _penv->GetNameId()%_bIsSelfCollisionChecker%pbody->GetName()%pbody->GetEnvironmentBodyIndex());
-        pinfo->Reset();
-        pinfo->_pbody = boost::const_pointer_cast<KinBody>(pbody);
-        // make sure that synchronization do occur !
-        pinfo->nLastStamp = pbody->GetUpdateStamp() - 1;
+    bool HasNamedGeometry(const KinBody &body, const std::string& groupname);
 
-        pinfo->vlinks.reserve(pbody->GetLinks().size());
-        FOREACHC(itlink, pbody->GetLinks()) {
-            const KinBody::LinkPtr& plink = *itlink;
-            boost::shared_ptr<FCLKinBodyInfo::LinkInfo> linkinfo(new FCLKinBodyInfo::LinkInfo(plink));
+    void SetGeometryGroup(const std::string& groupname);
 
+    const std::string& GetGeometryGroup() const;
 
-            fcl::AABB enclosingBV;
+    bool SetBodyGeometryGroup(KinBodyConstPtr pbody, const std::string& groupname);
 
-            // Glue code for a unified access to geometries
-            if(pinfo->_geometrygroup.size() > 0 && plink->GetGroupNumGeometries(pinfo->_geometrygroup) >= 0) {
-                const std::vector<KinBody::GeometryInfoPtr>& vgeometryinfos = plink->GetGeometriesFromGroup(pinfo->_geometrygroup);
-                FOREACH(itgeominfo, vgeometryinfos) {
-                    const KinBody::GeometryInfoPtr& pgeominfo = *itgeominfo;
-                    if( !pgeominfo ) {
-                        int igeominfo = itgeominfo - vgeometryinfos.begin();
-                        throw OpenRAVE::OpenRAVEException(str(boost::format("Failed to access geometry info %d for link %s:%s with geometrygroup %s")%igeominfo%plink->GetParent()->GetName()%plink->GetName()%pinfo->_geometrygroup), OpenRAVE::ORE_InvalidState);
-                    }
-                    const KinBody::GeometryInfo& geominfo = *pgeominfo;
-                    const CollisionGeometryPtr pfclgeom = _CreateFCLGeomFromGeometryInfo(_meshFactory, geominfo);
-
-                    if( !pfclgeom ) {
-                        continue;
-                    }
-                    pfclgeom->setUserData(nullptr);
-
-                    // We do not set the transformation here and leave it to _Synchronize
-                    CollisionObjectPtr pfclcoll = boost::make_shared<fcl::CollisionObject>(pfclgeom);
-                    pfclcoll->setUserData(linkinfo.get());
-                    linkinfo->vgeoms.push_back(TransformCollisionPair(geominfo.GetTransform(), pfclcoll));
-
-                    KinBody::Link::Geometry _tmpgeometry(boost::shared_ptr<KinBody::Link>(), geominfo);
-                    if( itgeominfo == vgeometryinfos.begin() ) {
-                        enclosingBV = ConvertAABBToFcl(_tmpgeometry.ComputeAABB(Transform()));
-                    }
-                    else {
-                        enclosingBV += ConvertAABBToFcl(_tmpgeometry.ComputeAABB(Transform()));
-                    }
-                }
-            }
-            else {
-                const std::vector<KinBody::Link::GeometryPtr> & vgeometries = plink->GetGeometries();
-                FOREACH(itgeom, vgeometries) {
-                    const KinBody::GeometryPtr& pgeom = *itgeom;
-                    const KinBody::GeometryInfo& geominfo = pgeom->GetInfo();
-                    const CollisionGeometryPtr pfclgeom = _CreateFCLGeomFromGeometryInfo(_meshFactory, geominfo);
-
-                    if( !pfclgeom ) {
-                        continue;
-                    }
-                    boost::shared_ptr<FCLKinBodyInfo::FCLGeometryInfo> pfclgeominfo(new FCLKinBodyInfo::FCLGeometryInfo(pgeom));
-                    pfclgeominfo->bodylinkgeomname = pbody->GetName() + "/" + plink->GetName() + "/" + pgeom->GetName();
-                    pfclgeom->setUserData(pfclgeominfo.get());
-                    // save the pointers
-                    linkinfo->vgeominfos.push_back(pfclgeominfo);
-
-                    // We do not set the transformation here and leave it to _Synchronize
-                    CollisionObjectPtr pfclcoll = boost::make_shared<fcl::CollisionObject>(pfclgeom);
-                    pfclcoll->setUserData(linkinfo.get());
-
-                    linkinfo->vgeoms.push_back(TransformCollisionPair(geominfo.GetTransform(), pfclcoll));
-
-                    KinBody::Link::Geometry _tmpgeometry(boost::shared_ptr<KinBody::Link>(), geominfo);
-                    if( itgeom == vgeometries.begin() ) {
-                        enclosingBV = ConvertAABBToFcl(_tmpgeometry.ComputeAABB(Transform()));
-                    }
-                    else {
-                        enclosingBV += ConvertAABBToFcl(_tmpgeometry.ComputeAABB(Transform()));
-                    }
-                }
-            }
-
-            if( linkinfo->vgeoms.size() == 0 ) {
-                RAVELOG_DEBUG_FORMAT("env=%s, Initializing body '%s' (index=%d) link '%s' with 0 geometries (env %d) (userdatakey %s)", _penv->GetNameId()%pbody->GetName()%pbody->GetEnvironmentBodyIndex()%plink->GetName()%_penv->GetId()%_userdatakey);
-            }
-            else {
-                CollisionGeometryPtr pfclgeomBV = std::make_shared<fcl::Box>(enclosingBV.max_ - enclosingBV.min_);
-                pfclgeomBV->setUserData(nullptr);
-                CollisionObjectPtr pfclcollBV = boost::make_shared<fcl::CollisionObject>(pfclgeomBV);
-                Transform trans(Vector(1,0,0,0),ConvertVectorFromFCL(0.5 * (enclosingBV.min_ + enclosingBV.max_)));
-                pfclcollBV->setUserData(linkinfo.get());
-                linkinfo->linkBV = std::make_pair(trans, pfclcollBV);
-            }
-
-            //link->nLastStamp = pinfo->nLastStamp;
-            linkinfo->bodylinkname = pbody->GetName() + "/" + plink->GetName();
-            pinfo->vlinks.push_back(linkinfo);
-#ifdef FCLRAVE_COLLISION_OBJECTS_STATISTICS
-            RAVELOG_DEBUG_FORMAT("FCLSPACECOLLISIONOBJECT|%s|%s", linkinfo->linkBV.second.get()%linkinfo->bodylinkname);
-#endif
-        }
-
-        pinfo->_geometrycallback = pbody->RegisterChangeCallback(KinBody::Prop_LinkGeometry, boost::bind(&FCLSpace::_ResetCurrentGeometryCallback,boost::bind(&OpenRAVE::utils::sptr_from<FCLSpace>, weak_space()),boost::weak_ptr<FCLKinBodyInfo>(pinfo)));
-        pinfo->_geometrygroupcallback = pbody->RegisterChangeCallback(KinBody::Prop_LinkGeometryGroup, boost::bind(&FCLSpace::_ResetGeometryGroupsCallback,boost::bind(&OpenRAVE::utils::sptr_from<FCLSpace>, weak_space()),boost::weak_ptr<FCLKinBodyInfo>(pinfo)));
-        pinfo->_linkenablecallback = pbody->RegisterChangeCallback(KinBody::Prop_LinkEnable, boost::bind(&FCLSpace::_ResetLinkEnableCallback, boost::bind(&OpenRAVE::utils::sptr_from<FCLSpace>, weak_space()), boost::weak_ptr<FCLKinBodyInfo>(pinfo)));
-        pinfo->_activeDOFsCallback = pbody->RegisterChangeCallback(KinBody::Prop_RobotActiveDOFs, boost::bind(&FCLSpace::_ResetActiveDOFsCallback, boost::bind(&OpenRAVE::utils::sptr_from<FCLSpace>, weak_space()), boost::weak_ptr<FCLKinBodyInfo>(pinfo)));
-
-        pinfo->_bodyAttachedCallback = pbody->RegisterChangeCallback(KinBody::Prop_BodyAttached, boost::bind(&FCLSpace::_ResetAttachedBodyCallback, boost::bind(&OpenRAVE::utils::sptr_from<FCLSpace>, weak_space()), boost::weak_ptr<FCLKinBodyInfo>(pinfo)));
-        pinfo->_bodyremovedcallback = pbody->RegisterChangeCallback(KinBody::Prop_BodyRemoved, boost::bind(&FCLSpace::RemoveUserData, boost::bind(&OpenRAVE::utils::sptr_from<FCLSpace>, weak_space()), boost::bind(&OpenRAVE::utils::sptr_from<const KinBody>, boost::weak_ptr<const KinBody>(pbody))));
-
-        const int envId = pbody->GetEnvironmentBodyIndex();
-        const int maxEnvId = _penv->GetMaxEnvironmentBodyIndex();
-        BOOST_ASSERT(envId != 0);
-        if( bSetToCurrentPInfo ) {
-            EnsureVectorSize(_currentpinfo, maxEnvId + 1);
-            _currentpinfo.at(envId) = pinfo;
-        }
-        //_cachedpinfo[pbody->GetEnvironmentBodyIndex()] what to do with the cache?
-        EnsureVectorSize(_vecInitializedBodies, maxEnvId + 1);
-        _vecInitializedBodies.at(envId) = pbody;
-
-        //Do I really need to synchronize anything at that point ?
-        _Synchronize(*pinfo, *pbody);
-
-        return pinfo;
-    }
-
-    bool HasNamedGeometry(const KinBody &body, const std::string& groupname) {
-        // The empty string corresponds to current geometries so all kinbodies have it
-        if( groupname.size() == 0 ) {
-            return true;
-        }
-        FOREACH(itlink, body.GetLinks()) {
-            if( (*itlink)->GetGroupNumGeometries(groupname) >= 0 ) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-
-    void SetGeometryGroup(const std::string& groupname)
-    {
-        // should always do this since bodies can have different geometry groups set
-        _geometrygroup = groupname;
-        for (const KinBodyConstPtr& pbody : _vecInitializedBodies) {
-            if (!pbody) {
-                continue;
-            }
-            SetBodyGeometryGroup(pbody, groupname);
-        }
-    }
-
-    const std::string& GetGeometryGroup() const
-    {
-        return _geometrygroup;
-    }
-
-
-    bool SetBodyGeometryGroup(KinBodyConstPtr pbody, const std::string& groupname) {
-        const KinBody& body = *pbody;
-
-        if (!HasNamedGeometry(body, groupname)) {
-            return false;
-        }
-
-        // Save the already existing FCLKinBodyInfoPtr for the old geometry group
-        FCLKinBodyInfoPtr& poldinfo = GetInfo(body);
-        if( poldinfo->_geometrygroup == groupname ) {
-            return true;
-        }
-
-        poldinfo->nGeometryUpdateStamp += 1;
-
-        const int maxBodyIndex = _penv->GetMaxEnvironmentBodyIndex();
-        EnsureVectorSize(_cachedpinfo, maxBodyIndex + 1);
-
-        const int bodyIndex = body.GetEnvironmentBodyIndex();
-        OPENRAVE_ASSERT_OP_FORMAT(bodyIndex, !=, 0, "env=%s, body %s", _penv->GetNameId()%body.GetName(), OpenRAVE::ORE_InvalidState);
-        std::map< std::string, FCLKinBodyInfoPtr >& cache = _cachedpinfo.at(bodyIndex);
-        cache[poldinfo->_geometrygroup] = poldinfo;
-
-        FCLKinBodyInfoPtr& pinfo = cache[groupname];
-        if(!pinfo) {
-            RAVELOG_VERBOSE_FORMAT("FCLSpace : creating geometry %s for kinbody %s (id = %d) (env = %d)", groupname%body.GetName()%bodyIndex%_penv->GetId());
-            pinfo.reset(new FCLKinBodyInfo);
-            pinfo->_geometrygroup = groupname;
-            InitKinBody(pbody, pinfo);
-        }
-        else {
-            RAVELOG_VERBOSE_FORMAT("env=%s, switching to geometry %s for kinbody %s (id = %d)", _penv->GetNameId()%groupname%body.GetName()%bodyIndex);
-            // Set the current info to use the FCLKinBodyInfoPtr associated to groupname
-            EnsureVectorSize(_currentpinfo, maxBodyIndex + 1);
-            _currentpinfo.at(bodyIndex) = pinfo;
-
-            // Revoke the information inside the cache so that a potentially outdated object does not survive
-            cache.erase(groupname);
-        }
-
-        return true;
-    }
-
-    const std::string& GetBodyGeometryGroup(const KinBody &body) const {
-        static const std::string empty;
-        const FCLKinBodyInfoPtr& pinfo = GetInfo(body);
-        if( !!pinfo ) {
-            return pinfo->_geometrygroup;
-        } else {
-            return empty;
-        }
-    }
+    const std::string& GetBodyGeometryGroup(const KinBody &body) const;
 
     // Set the current bvhRepresentation and reinitializes all the KinbodyInfo if needed
-    void SetBVHRepresentation(std::string const &type)
-    {
-        if( type == _bvhRepresentation ) {
-            return;
-        }
+    void SetBVHRepresentation(std::string const &type);
 
-        if (type == "AABB") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL<fcl::AABB>;
-        } else if (type == "OBB") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL<fcl::OBB>;
-        } else if (type == "RSS") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL<fcl::RSS>;
-        } else if (type == "OBBRSS") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL<fcl::OBBRSS>;
-        } else if (type == "kDOP16") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL< fcl::KDOP<16> >;
-        } else if (type == "kDOP18") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL< fcl::KDOP<18> >;
-        } else if (type == "kDOP24") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL< fcl::KDOP<24> >;
-        } else if (type == "kIOS") {
-            _bvhRepresentation = type;
-            _meshFactory = &ConvertMeshToFCL<fcl::kIOS>;
-        } else {
-            RAVELOG_WARN(str(boost::format("Unknown BVH representation '%s', keeping '%s' representation") % type % _bvhRepresentation));
-            return;
-        }
+    std::string const& GetBVHRepresentation() const;
 
-        // reinitialize all the FCLKinBodyInfo
+    void Synchronize();
 
-        for (const KinBodyConstPtr& pbody : _vecInitializedBodies) {
-            if (!pbody) {
-                continue;
-            }
-            const KinBody& body = *pbody;
-            FCLKinBodyInfoPtr& pinfo = GetInfo(body);
-            pinfo->nGeometryUpdateStamp++;
-            InitKinBody(pbody, pinfo);
-        }
-        _cachedpinfo.clear();
-    }
+    void Synchronize(const KinBody &body);
 
-    std::string const& GetBVHRepresentation() const {
-        return _bvhRepresentation;
-    }
+    void SynchronizeWithAttached(const KinBody &body);
 
+    FCLKinBodyInfoPtr& GetInfo(const KinBody &body);
 
-    void Synchronize()
-    {
-        // We synchronize only the initialized bodies, which differs from oderave
-        for (const KinBodyConstPtr& pbody : _vecInitializedBodies) {
-            if (!pbody) {
-                continue;
-            }
-            Synchronize(*pbody);
-        }
-    }
+    const FCLKinBodyInfoPtr& GetInfo(const KinBody &body) const;
 
-    void Synchronize(const KinBody &body)
-    {
-        FCLKinBodyInfoPtr& pinfo = GetInfo(body);
-        if( !pinfo ) {
-            return;
-        }
-        // expensive, comment out for now
-        //BOOST_ASSERT( pinfo->_pbody.lock().get() == &body);
-        _Synchronize(*pinfo, body);
-    }
-
-    void SynchronizeWithAttached(const KinBody &body)
-    {
-        if( body.HasAttached() ) {
-            std::vector<int>& vecAttachedEnvBodyIndices = _vecAttachedEnvBodyIndicesCache;
-            body.GetAttachedEnvironmentBodyIndices(vecAttachedEnvBodyIndices);
-            std::vector<KinBodyPtr>& attachedBodies = _vecAttachedBodiesCache;
-            _penv->GetBodiesFromEnvironmentBodyIndices(vecAttachedEnvBodyIndices, attachedBodies);
-            for (const KinBodyPtr& pattachedBody : attachedBodies) {
-                if (!!pattachedBody) {
-                    Synchronize(*pattachedBody);
-                }
-            }
-        }
-        else {
-            Synchronize(body);
-        }
-    }
-
-    inline FCLKinBodyInfoPtr& GetInfo(const KinBody &body)
-    {
-        int envId = body.GetEnvironmentBodyIndex();
-        if ( envId <= 0 ) {
-            RAVELOG_WARN_FORMAT("env=%s, body %s has invalid environment id %d", _penv->GetNameId()%body.GetName()%envId);
-        }
-        else {
-            if (envId < (int) _currentpinfo.size()) {
-                return _currentpinfo.at(envId);
-            }
-        }
-
-        return _currentpinfo.at(0); // invalid
-    }
-
-    inline const FCLKinBodyInfoPtr& GetInfo(const KinBody &body) const
-    {
-        int envId = body.GetEnvironmentBodyIndex();
-        if ( envId <= 0 ) {
-            RAVELOG_WARN_FORMAT("env=%s, body %s has invalid environment id %d", _penv->GetNameId()%body.GetName()%envId);
-        }
-        else {
-            if (envId < (int) _currentpinfo.size()) {
-                return _currentpinfo.at(envId);
-            }
-        }
-
-        return _currentpinfo.at(0);
-    }
-
-
-    void RemoveUserData(KinBodyConstPtr pbody) {
-        if( !!pbody ) {
-            RAVELOG_VERBOSE(str(boost::format("FCL User data removed from env %d (userdatakey %s) : %s") % _penv->GetId() % _userdatakey % pbody->GetName()));
-            const int envId = pbody->GetEnvironmentBodyIndex();
-            if (envId < (int) _vecInitializedBodies.size()) {
-                _vecInitializedBodies.at(envId).reset();
-            }
-            FCLKinBodyInfoPtr& pinfo = GetInfo(*pbody);
-            if( !!pinfo ) {
-                pinfo->Reset();
-            }
-
-            if( envId == 0 ) {
-                RAVELOG_WARN_FORMAT("env=%s, body '%s' has bodyIndex=0, so not adding to the environment!", _penv->GetNameId()%pbody->GetName());
-            }
-
-            if (envId < (int) _currentpinfo.size()) {
-                _currentpinfo.at(envId).reset();
-                //RAVELOG_INFO_FORMAT("erased %d but didn't pop back, size is %d", envId%_currentpinfo.size());
-            }
-            if (envId < (int) _cachedpinfo.size()) {
-                _cachedpinfo.at(envId).clear();
-            }
-        }
-    }
-
+    void RemoveUserData(KinBodyConstPtr pbody);
 
     /// \brief returns bodies initialized by this space. Note that some entries are null pointer.
     const std::vector<KinBodyConstPtr>& GetEnvBodies() const {
@@ -651,140 +260,12 @@ public:
         return _penv->GetId();
     }
 private:
-    static void _AddGeomInfoToBVHSubmodel(fcl::BVHModel<fcl::OBB>& model, KinBody::GeometryInfo const &info)
-    {
-        const OpenRAVE::TriMesh& mesh = info._meshcollision;
-        if (mesh.vertices.empty() || mesh.indices.empty()) {
-            return;
-        }
-
-        OPENRAVE_ASSERT_OP(mesh.indices.size() % 3, ==, 0);
-        size_t const num_points = mesh.vertices.size();
-        size_t const num_triangles = mesh.indices.size() / 3;
-
-        std::vector<fcl::Vec3f> fcl_points(num_points);
-        for (size_t ipoint = 0; ipoint < num_points; ++ipoint) {
-            Vector v = info.GetTransform()*mesh.vertices[ipoint];
-            fcl_points[ipoint] = fcl::Vec3f(v.x, v.y, v.z);
-        }
-
-        std::vector<fcl::Triangle> fcl_triangles(num_triangles);
-        for (size_t itri = 0; itri < num_triangles; ++itri) {
-            int const *const tri_indices = &mesh.indices[3 * itri];
-            fcl_triangles[itri] = fcl::Triangle(tri_indices[0], tri_indices[1], tri_indices[2]);
-        }
-        model.addSubModel(fcl_points, fcl_triangles);
-    }
-
-    static TransformCollisionPair _CreateTransformCollisionPairFromOBB(fcl::OBB const &bv) {
-        CollisionGeometryPtr pbvGeom = make_shared<fcl::Box>(bv.extent[0]*2.0f, bv.extent[1]*2.0f, bv.extent[2]*2.0f);
-        CollisionObjectPtr pbvColl = boost::make_shared<fcl::CollisionObject>(pbvGeom);
-        fcl::Quaternion3f fclBvRot;
-        fclBvRot.fromAxes(bv.axis);
-        Vector bvRotation = ConvertQuaternionFromFCL(fclBvRot);
-        Vector bvTranslation = ConvertVectorFromFCL(bv.center());
-
-        return std::make_pair(Transform(bvRotation, bvTranslation), pbvColl);
-    }
 
     // what about the tests on non-zero size (eg. box extents) ?
-    static CollisionGeometryPtr _CreateFCLGeomFromGeometryInfo(const MeshFactory &mesh_factory, const KinBody::GeometryInfo &info)
-    {
-        switch(info._type) {
-
-        case OpenRAVE::GT_None:
-            return CollisionGeometryPtr();
-
-        case OpenRAVE::GT_Box:
-            return make_shared<fcl::Box>(info._vGeomData.x*2.0f,info._vGeomData.y*2.0f,info._vGeomData.z*2.0f);
-
-        case OpenRAVE::GT_Sphere:
-            return make_shared<fcl::Sphere>(info._vGeomData.x);
-
-        case OpenRAVE::GT_Cylinder:
-            return make_shared<fcl::Cylinder>(info._vGeomData.x, info._vGeomData.y);
-
-        case OpenRAVE::GT_Container:
-        case OpenRAVE::GT_TriMesh:
-        case OpenRAVE::GT_Cage:
-        case OpenRAVE::GT_CalibrationBoard: // calibration board is box-shaped but has z-offset. so have to use trimesh.
-        {
-            const OpenRAVE::TriMesh& mesh = info._meshcollision;
-            if (mesh.vertices.empty() || mesh.indices.empty()) {
-                return CollisionGeometryPtr();
-            }
-
-            OPENRAVE_ASSERT_OP(mesh.indices.size() % 3, ==, 0);
-            size_t const num_points = mesh.vertices.size();
-            size_t const num_triangles = mesh.indices.size() / 3;
-
-            std::vector<fcl::Vec3f> fcl_points(num_points);
-            for (size_t ipoint = 0; ipoint < num_points; ++ipoint) {
-                Vector v = mesh.vertices[ipoint];
-                fcl_points[ipoint] = fcl::Vec3f(v.x, v.y, v.z);
-            }
-
-            std::vector<fcl::Triangle> fcl_triangles(num_triangles);
-            for (size_t itri = 0; itri < num_triangles; ++itri) {
-                int const *const tri_indices = &mesh.indices[3 * itri];
-                fcl_triangles[itri] = fcl::Triangle(tri_indices[0], tri_indices[1], tri_indices[2]);
-            }
-
-            return mesh_factory(fcl_points, fcl_triangles);
-        }
-
-        default:
-            RAVELOG_WARN(str(boost::format("FCL doesn't support geom type %d")%info._type));
-            return CollisionGeometryPtr();
-        }
-    }
+    CollisionGeometryPtr _CreateFCLGeomFromGeometryInfo(const KinBody::GeometryInfo &info);
 
     /// \brief pass in info.GetBody() as a reference to avoid dereferencing the weak pointer in FCLKinBodyInfo
-    void _Synchronize(FCLKinBodyInfo& info, const KinBody& body)
-    {
-        //KinBodyPtr pbody = info.GetBody();
-        if( info.nLastStamp != body.GetUpdateStamp()) {
-            info.nLastStamp = body.GetUpdateStamp();
-            if( body.GetLinks().size() != info.vlinks.size() ) {
-                throw OpenRAVE::OpenRAVEException(str(boost::format("env=%s, the current number of links in body '%s' are %d, and are not the same as the number cached links %d")%_penv->GetNameId()%body.GetName()%body.GetLinks().size()%info.vlinks.size()), OpenRAVE::ORE_InvalidState);
-            }
-            CollisionObjectPtr pcoll;
-            for(size_t i = 0; i < body.GetLinks().size(); ++i) {
-                const FCLSpace::FCLKinBodyInfo::LinkInfo& linkInfo = *info.vlinks[i];
-                pcoll = linkInfo.linkBV.second;
-                if( !pcoll ) {
-                    continue;
-                }
-                const Transform& linkTransform = body.GetLinks()[i]->GetTransform();
-                Transform pose = linkTransform * linkInfo.linkBV.first;
-                fcl::Vec3f newPosition = ConvertVectorToFCL(pose.trans);
-                fcl::Quaternion3f newOrientation = ConvertQuaternionToFCL(pose.rot);
-
-                pcoll->setTranslation(newPosition);
-                pcoll->setQuatRotation(newOrientation);
-                // Do not forget to recompute the AABB otherwise getAABB won't give an up to date AABB
-                pcoll->computeAABB();
-
-                //info.vlinks[i]->nLastStamp = info.nLastStamp;
-                for (const TransformCollisionPair& pgeom : linkInfo.vgeoms) {
-                    fcl::CollisionObject& coll = *pgeom.second;
-                    Transform pose1 = linkTransform * pgeom.first;
-                    fcl::Vec3f newPosition1 = ConvertVectorToFCL(pose1.trans);
-                    fcl::Quaternion3f newOrientation1 = ConvertQuaternionToFCL(pose1.rot);
-
-                    coll.setTranslation(newPosition1);
-                    coll.setQuatRotation(newOrientation1);
-                    // Do not forget to recompute the AABB otherwise getAABB won't give an up to date AABB
-                    coll.computeAABB();
-                }
-            }
-
-            // Does this have any use ?
-            // if( !!_synccallback ) {
-            //     _synccallback(pinfo);
-            // }
-        }
-    }
+    void _Synchronize(FCLKinBodyInfo& info, const KinBody& body);
 
     /// \brief controls whether the kinbody info is removed during the destructor
     class FCLKinBodyInfoRemover
@@ -807,46 +288,9 @@ private:
         boost::function<void()> _fn;
         bool _bDoRemove;
     };
-    void _ResetCurrentGeometryCallback(boost::weak_ptr<FCLKinBodyInfo> _pinfo)
-    {
-        FCLKinBodyInfoPtr pinfo = _pinfo.lock();
-        KinBodyPtr pbody = pinfo->GetBody();
-        const int bodyIndex = pbody->GetEnvironmentBodyIndex();
-        if (0 < bodyIndex && bodyIndex < (int)_currentpinfo.size()) {
-            const FCLKinBodyInfoPtr& pcurrentinfo = _currentpinfo.at(bodyIndex);
+    void _ResetCurrentGeometryCallback(boost::weak_ptr<FCLKinBodyInfo> _pinfo);
 
-            if( !!pinfo && pinfo == pcurrentinfo ) {//pinfo->_geometrygroup.size() == 0 ) {
-                // pinfo is current set to the current one, so should InitKinBody into _currentpinfo
-                //RAVELOG_VERBOSE_FORMAT("env=%d, resetting current geometry for kinbody %s nGeometryUpdateStamp=%d, (key %s, self=%d)", _penv->GetId()%pbody->GetName()%pinfo->nGeometryUpdateStamp%_userdatakey%_bIsSelfCollisionChecker);
-                pinfo->nGeometryUpdateStamp++;
-                FCLKinBodyInfoRemover remover(boost::bind(&FCLSpace::RemoveUserData, this, pbody)); // protect
-                InitKinBody(pbody, pinfo, false);
-                remover.ResetRemove(); // succeeded
-            }
-            //_cachedpinfo[pbody->GetEnvironmentBodyIndex()].erase(std::string());
-        }
-    }
-
-    void _ResetGeometryGroupsCallback(boost::weak_ptr<FCLKinBodyInfo> _pinfo)
-    {
-        FCLKinBodyInfoPtr pinfo = _pinfo.lock();
-        KinBodyPtr pbody = pinfo->GetBody();
-
-        //FCLKinBodyInfoPtr pcurrentinfo = _currentpinfo.at(pbody->GetEnvironmentBodyIndex());
-
-        if( !!pinfo ) {// && pinfo->_geometrygroup.size() > 0 ) {
-            //RAVELOG_VERBOSE_FORMAT("env=%d, resetting geometry groups for kinbody %s, nGeometryUpdateStamp=%d (key %s, self=%d)", _penv->GetId()%pbody->GetName()%pinfo->nGeometryUpdateStamp%_userdatakey%_bIsSelfCollisionChecker);
-            pinfo->nGeometryUpdateStamp++;
-            FCLKinBodyInfoRemover remover(boost::bind(&FCLSpace::RemoveUserData, this, pbody)); // protect
-            InitKinBody(pbody, pinfo, false);
-            remover.ResetRemove(); // succeeded
-        }
-//        FCLKinBodyInfoPtr pinfoCurrentGeometry = _cachedpinfo[pbody->GetEnvironmentBodyIndex()][std::string()];
-//        _cachedpinfo.erase(pbody->GetEnvironmentBodyIndex());
-//        if( !!pinfoCurrentGeometry ) {
-//            _cachedpinfo[pbody->GetEnvironmentBodyIndex()][std::string()] = pinfoCurrentGeometry;
-//        }
-    }
+    void _ResetGeometryGroupsCallback(boost::weak_ptr<FCLKinBodyInfo> _pinfo);
 
     void _ResetLinkEnableCallback(boost::weak_ptr<FCLKinBodyInfo> _pinfo) {
         FCLKinBodyInfoPtr pinfo = _pinfo.lock();
@@ -868,7 +312,6 @@ private:
             pinfo->nAttachedBodiesUpdateStamp++;
         }
     }
-
 
     EnvironmentBasePtr _penv;
     std::string _userdatakey;
